consolidated canvas capturer code
