+++ lib/capture-canvas.js
@@ -30,10 +30,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-const { makeFilePathConverter, writeFile } = require('./utils.js');
-const path = require('path');
-
-const defaultCanvasMode = 'png';
+const makeCanvasCapturer = require('./make-canvas-capturer');
 
 const canvasToBuffer = function (page, canvasSelector, type) {
   return page.evaluate(function (canvasSelector, type) {
@@ -45,51 +42,4 @@ const canvasToBuffer = function (page, canvasSelector, type) {
   });
 };
 
-module.exports = function (config) {
-  var page = config.page;
-  var log = config.log;
-  var frameProcessor = config.frameProcessor;
-  var filePathConverter = makeFilePathConverter(config);
-  var canvasMode = config.canvasMode;
-  var canvasSelector = config.selector || 'canvas';
-  var pendingWritePromises = [];
-  var waitForWriting = false;
-  return {
-    capture: function (sameConfig, frameCount, framesToCapture) {
-      var filePath = filePathConverter(frameCount, framesToCapture);
-      if (filePath && typeof canvasMode !== 'string') {
-        canvasMode = path.extname(filePath).substring(1);
-      }
-      if (typeof canvasMode !== 'string') {
-        canvasMode = defaultCanvasMode;
-      }
-      if (!canvasMode.startsWith('image/')) {
-        canvasMode = 'image/' + canvasMode;
-      }
-      log('Capturing Frame ' + frameCount + (filePath ? ' to ' + filePath : '') + '...');
-      var p = canvasToBuffer(page, canvasSelector, canvasMode);
-      if (filePath) {
-        p = p.then(function (buffer) {
-          var writePromise = writeFile(filePath, buffer);
-          if (waitForWriting) {
-            return writePromise.then(function () {
-              return buffer;
-            });
-          } else {
-            pendingWritePromises.push(writePromise);
-            return buffer;
-          }
-        });
-      }
-      if (frameProcessor) {
-        p = p.then(function (buffer) {
-          return frameProcessor(buffer, frameCount, framesToCapture);
-        });
-      }
-      return p;
-    },
-    afterCapture: function () {
-      return Promise.all(pendingWritePromises);
-    }
-  };
-};
+module.exports = makeCanvasCapturer(canvasToBuffer);
