+++ lib/immediate-canvas-handler.js
@@ -31,31 +31,22 @@
  */
 
 const { overwriteTime } = require('./overwrite-time');
-const { makeFilePathConverter, writeFile } = require('./utils.js');
-const path = require('path');
+const makeCanvasCapturer = require('./make-canvas-capturer');
 
-const defaultCanvasMode = 'png';
+const canvasCapturer = makeCanvasCapturer(function (page) {
+  return page.evaluate(function () {
+    return window._timesnap_canvasData;
+  }).then(function (dataUrl) {
+    var data = dataUrl.slice(dataUrl.indexOf(',') + 1);
+    return new Buffer(data, 'base64');
+  });
+});
 
 module.exports = function (config) {
+  var capturer = canvasCapturer(config);
+  var canvasMode = capturer.canvasMode;
+  var canvasSelector = capturer.canvasSelector;
   var page = config.page;
-  var log = config.log;
-  var frameProcessor = config.frameProcessor;
-  var filePathConverter = makeFilePathConverter(config);
-  var canvasMode = config.canvasMode || null;
-  var canvasSelector = config.selector || 'canvas';
-  var pendingWritePromises = [];
-  var waitForWriting = false;
-
-  var filePath = filePathConverter(1, 100); // just get a sample file name
-  if (filePath && typeof canvasMode !== 'string') {
-    canvasMode = path.extname(filePath).substring(1);
-  }
-  if (typeof canvasMode !== 'string') {
-    canvasMode = defaultCanvasMode;
-  }
-  if (!canvasMode.startsWith('image/')) {
-    canvasMode = 'image/' + canvasMode;
-  }
 
   const goToTime = function (browserFrames, time) {
     // Goes to a certain time. Can't go backwards
@@ -67,38 +58,8 @@ module.exports = function (config) {
   };
 
   return {
-    capture: function (sameConfig, frameCount, framesToCapture) {
-      filePath = filePathConverter(frameCount, framesToCapture);
-      log('Capturing Frame ' + frameCount + (filePath ? ' to ' + filePath : '') + '...');
-      var p = page.evaluate(function () {
-        return window._timesnap_canvasData;
-      }).then(function (dataUrl) {
-        var data = dataUrl.slice(dataUrl.indexOf(',') + 1);
-        return new Buffer(data, 'base64');
-      });
-      if (filePath) {
-        p = p.then(function (buffer) {
-          var writePromise = writeFile(filePath, buffer);
-          if (waitForWriting) {
-            return writePromise.then(function () {
-              return buffer;
-            });
-          } else {
-            pendingWritePromises.push(writePromise);
-            return buffer;
-          }
-        });
-      }
-      if (frameProcessor) {
-        p = p.then(function (buffer) {
-          return frameProcessor(buffer, frameCount, framesToCapture);
-        });
-      }
-      return p;
-    },
-    afterCapture: function () {
-      return Promise.all(pendingWritePromises);
-    },
+    capture: capturer.capture,
+    afterCapture: capturer.afterCapture,
     overwriteTime,
     goToTime
   };
